# colblk: don't use UnsafeRawSlice to store PrefixBytesBuilder offsets