--- a/sstable/colblk/raw_bytes.go
+++ b/sstable/colblk/raw_bytes.go
@@ -11,7 +11,6 @@ import (
 	"unsafe"
 
 	"github.com/cockroachdb/pebble/internal/binfmt"
-	"github.com/cockroachdb/pebble/internal/invariants"
 	"github.com/cockroachdb/pebble/internal/treeprinter"
 )
 
@@ -221,12 +220,4 @@ func (b *RawBytesBuilder) Size(rows int, offset uint32) uint32 {
 // WriteDebug implements Encoder.
 func (b *RawBytesBuilder) WriteDebug(w io.Writer, rows int) {
 	fmt.Fprintf(w, "bytes: %d rows set; %d bytes in data", b.rows, len(b.data))
-}
-
-//gcassert:inline
-func unsafeGetUint64(slice []uint64, idx int) uint64 {
-	if invariants.Enabled {
-		_ = slice[idx]
-	}
-	return *(*uint64)(unsafe.Add(unsafe.Pointer(unsafe.SliceData(slice)), uintptr(idx)<<3))
 }