#### colblk: add uintsEncoder helper

This replaces the use of `UnsafeRawSlice` for writing out integers.
This separation will help with cleanly and efficiently adding big
endian support.

#### colblk: don't use UnsafeRawSlice to store PrefixBytesBuilder offsets

We can use a simple slice, along with some helpers that avoid bounds
checking. UnsafeRawSlice can now only be used to read values.

---

No differences in ColBlockWriter benchmarks:
```
name                                                                                                    old time/op    new time/op    delta
CockroachDataColBlockWriter/AlphaLen=8,Prefix=8,Shared=4,KeysPerPrefix=4,Logical=10,ValueLen=8-10         63.5µs ± 1%    63.1µs ± 3%   ~     (p=0.113 n=9+10)
CockroachDataColBlockWriter/AlphaLen=8,Prefix=128,Shared=64,KeysPerPrefix=4,Logical=50,ValueLen=128-10    15.3µs ± 2%    15.1µs ± 4%   ~     (p=0.393 n=10+10)
CockroachDataColBlockWriter/AlphaLen=26,Prefix=1024,Shared=512,KeysPerPrefix=1,ValueLen=1024-10           3.12µs ± 6%    3.13µs ± 5%   ~     (p=0.869 n=10+10)

name                                                                                                    old alloc/op   new alloc/op   delta
CockroachDataColBlockWriter/AlphaLen=8,Prefix=8,Shared=4,KeysPerPrefix=4,Logical=10,ValueLen=8-10          9.00B ± 0%     9.00B ± 0%   ~     (all equal)
CockroachDataColBlockWriter/AlphaLen=8,Prefix=128,Shared=64,KeysPerPrefix=4,Logical=50,ValueLen=128-10     2.00B ± 0%     2.00B ± 0%   ~     (all equal)
CockroachDataColBlockWriter/AlphaLen=26,Prefix=1024,Shared=512,KeysPerPrefix=1,ValueLen=1024-10            0.00B          0.00B        ~     (all equal)

name                                                                                                    old allocs/op  new allocs/op  delta
CockroachDataColBlockWriter/AlphaLen=8,Prefix=8,Shared=4,KeysPerPrefix=4,Logical=10,ValueLen=8-10           0.00           0.00        ~     (all equal)
CockroachDataColBlockWriter/AlphaLen=8,Prefix=128,Shared=64,KeysPerPrefix=4,Logical=50,ValueLen=128-10      0.00           0.00        ~     (all equal)
CockroachDataColBlockWriter/AlphaLen=26,Prefix=1024,Shared=512,KeysPerPrefix=1,ValueLen=1024-10             0.00           0.00        ~     (all equal)
```